Restructured Redact files
